Extract basicscroll in our codebase
There were some issues with basicscroll, first of all, it was using
imports, which didn't mix well with Gatsby's webpack babel
configuration.

Second of all, the module was calling a `loop` procedure on module
loading, which caused problems with server side rendering.

Now the basicscroll module will only start the loop as soon as `create`
was called for the first time. This gives us the possibility to handle
the module load gracefully without interfering with the SSR process.