Auto calculate from - to Timetable times
With this commit, we are not required to
handwrite our time-schedule times by hand...

Entries may provide a duration, which will be used to calculate
the fromTime from a previous entry. Entries represent calculated
timetable elements, which are calculated from a step value.

See calcStep / calcTimetable on how the entries are calculated from
a start value and a list of steps.